[UPDATE] salary column type to integer.

+++ database/migrations/2024_06_02_201856_create_offers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->bigInteger('salary');
+            $table->integer('salary');
             $table->integer('hours');
             $table->timestamp('date')->nullable();
             $table->string('requirements');
